reorg: put evaluation in "ShapeCollection"
