+++ Applications/shapeworks/CMakeLists.txt
@@ -4,14 +4,12 @@ set(shapeworks_sources
   Executable.cpp
   Command.cpp
   Commands.cpp
-  Analyze/EvaluateSSM.cpp
   )
 set(shapeworks_headers
   Executable.h
   Command.h
   Commands.h
   SharedCommandData.h
-  Analyze/EvaluateSSM.h
   )
 add_executable(shapeworks
   ${shapeworks_sources}
@@ -23,6 +21,7 @@ target_link_libraries(shapeworks
   Utils
   Image
   Mesh
+  Analyze
   ${ITK_LIBRARIES}
   )
